cloud_client seperate tests
