+++ otter/test/cloud_client/test_init.py
@@ -28,9 +28,9 @@ from otter.auth import Authenticate, InvalidateToken
 from otter.cloud_client import (
     CLBDeletedError,
     CLBDuplicateNodesError,
+    CLBImmutableError,
     CLBNodeLimitError,
     CLBNotActiveError,
-    CLBImmutableError,
     CLBRateLimitError,
     CreateServerConfigurationError,
     CreateServerOverQuoteError,
@@ -62,7 +62,6 @@ from otter.cloud_client import (
     list_servers_details_page,
     list_stacks_all,
     perform_tenant_scope,
-    publish_to_cloudfeeds,
     remove_clb_nodes,
     service_request,
     set_nova_metadata_item,
@@ -1378,38 +1377,6 @@ class NovaClientTests(SynchronousTestCase):
              lambda _: raise_(NovaComputeFaultError('error')))
         ]
         self.assertRaises(NovaComputeFaultError, perform_sequence, seq, eff)
-
-
-class CloudFeedsTests(SynchronousTestCase):
-    """
-    Tests for cloud feed functions.
-    """
-    def test_publish_to_cloudfeeds(self):
-        """
-        Publish an event to cloudfeeds.  Successfully handle non-JSON data.
-        """
-        _log = object()
-        eff = publish_to_cloudfeeds({'event': 'stuff'}, log=_log)
-        expected = service_request(
-            ServiceType.CLOUD_FEEDS, 'POST',
-            'autoscale/events',
-            headers={'content-type': ['application/vnd.rackspace.atom+json']},
-            data={'event': 'stuff'}, log=_log, success_pred=has_code(201),
-            json_response=False)
-
-        # success
-        dispatcher = EQFDispatcher([(
-            expected.intent,
-            service_request_eqf(stub_pure_response('<this is xml>', 201)))])
-        resp, body = sync_perform(dispatcher, eff)
-        self.assertEqual(body, '<this is xml>')
-
-        # Add regression test that 202 should be an API error because this
-        # is a bug in CF
-        dispatcher = EQFDispatcher([(
-            expected.intent,
-            service_request_eqf(stub_pure_response('<this is xml>', 202)))])
-        self.assertRaises(APIError, sync_perform, dispatcher, eff)
 
 
 class CloudOrchestrationTests(SynchronousTestCase):
